pyproject.toml
[project]
name = "flowa"
version = "0.1.0"
description = "Variant literature assessment pipeline with AI extraction"
readme = "README.md"
requires-python = "==3.13.*"
license = { file = "LICENSE" }
dependencies = [
    "defusedxml",
    "fsspec",
    "gcsfs",  # GCS support for fsspec
    "groundmark==0.2.0",
    "httpx",
    "pydantic-ai-slim[anthropic,bedrock,google,openai]>=1.69.0",
    "pydantic-settings",
    "pypdf",
    "s3fs",   # S3/MinIO support for fsspec
    "tenacity",
    "typer",
]

[dependency-groups]
dev = ["ruff", "mypy", "pytest", "types-defusedxml"]

[project.scripts]
flowa = "flowa.cli:app"

[tool.mypy]
python_version = "3.13"
disable_error_code = ["import-untyped"]

[tool.ruff]
target-version = "py313"
line-length = 120
src = ["src"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade - modernize Python code
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # Do not perform function call in argument defaults - false positive with Typer
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
